third_party/blink/web_tests/http/tests/loading/preload-image-srcset-duplicate-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - BeginNavigation request to 'http://127.0.0.1:8000/loading/preload-image-srcset-duplicate.html', http method GET
main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
PASS internals.isPreloaded('resources/preload-test.jpg'); is false
PASS internals.isPreloaded('resources/preload-test-duplicate.jpg'); is false
PASS internals.isPreloaded('resources/base-image1.png'); is true
PASS internals.isPreloaded('resources/base-image2.png'); is false
PASS internals.isPreloaded('resources/base-image3.png'); is false
PASS internals.isPreloaded('resources/dup-image1.png'); is false
PASS internals.isPreloaded('resources/dup-image2.png'); is false
PASS internals.isPreloaded('resources/dup-image3.png'); is false
PASS internals.isPreloaded('resources/image2.png'); is true
PASS successfullyParsed is true

TEST COMPLETE

This test passes if the fallback image (defined in the src attribute) is not preloaded, the 1x version (defined in srcset) is, and none of the duplicate resources is preloaded.
The second image is here to make sure isPreloaded is working properly
 
